Split into two packages
Until now a conda installation of silx resulted in an installation
with almost all required and optional dependencies. This commit splits
the package into silx and silx-base, with the latter containing all
actual Python code and binaries, but dragging in only the minimal
dependencies necessary to run silx. silx drags in everything, just like
before.